move DriverTestCase to extensions
