+++ src/qcodes/tests/drivers/test_Agilent_E8257D.py
@@ -1,5 +1,5 @@
+from qcodes.extensions import DriverTestCase
 from qcodes.instrument_drivers.agilent import AgilentE8257D
-from qcodes.tests.driver_test_case import DriverTestCase
 
 
 class TestAgilentE8257D(DriverTestCase):
